requirements.txt
requests>=2.31.0
aiohttp>=3.8.5
beautifulsoup4>=4.12.2
lxml>=4.9.3
dnspython>=2.4.2
python-whois>=0.8.0
shodan>=1.30.0
censys>=2.2.0
virustotal-python>=1.0.2
waybackpy>=0.4.0
click>=8.1.7
colorama>=0.4.6
tabulate>=0.9.0
jinja2>=3.1.2
pyyaml>=6.0.1
tqdm>=4.66.1
rich>=13.7.0
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
